fix: blog CRUD
- Refatorado get e update para update_all
- Refatorado get e delete para delete_all.
